move parking space lease endpoints to lease-service
